requirements.txt
pydobot2==0.1.0
pyserial==3.4
